+++ Models/Susy/SSGSSVertex.cc
@@ -21,13 +21,13 @@ using namespace ThePEG::Helicity;
 using namespace Herwig;
 
 SSGSSVertex::SSGSSVertex() : _couplast(0.),_q2last(0.*GeV2) {
-  vector<int> first,second,third;
-  for(int ix=1000001;ix<1000007;++ix) {
+  vector<long> first,second,third;
+  for(long ix=1000001;ix<1000007;++ix) {
     first.push_back(21);
     second.push_back(ix);
     third.push_back(-ix);
   }
-  for(int ix=2000001;ix<2000007;++ix) {
+  for(long ix=2000001;ix<2000007;++ix) {
     first.push_back(21);
     second.push_back(ix);
     third.push_back(-ix);
